2x speedup by using C rand()

+++ src/textures/image_texture.cpp
@@ -3,7 +3,6 @@
 #include <string>
 
 #include <glm/glm.hpp>
-#include <glm/gtc/random.hpp>
 
 
 namespace Raytracer
@@ -19,8 +18,8 @@ ImageTexture::ImageTexture(const std::string& path) noexcept
 glm::vec3 ImageTexture::colourAt(float u, float v, const glm::vec3& p) const
 {
     // convert to image coordinates
-    float x = static_cast<int>(glm::clamp(u, 0.0f, 0.9999f) * m_image->width());
-    float y = static_cast<int>(glm::clamp(v, 0.0f, 0.9999f) * m_image->height());
+    int x = static_cast<int>(glm::clamp(u, 0.0f, 0.9999f) * m_image->width());
+    int y = static_cast<int>(glm::clamp(v, 0.0f, 0.9999f) * m_image->height());
 
     return m_image->get(x, y);
 }
